help webpage 2.0 sound-search
Given the lack of use, I have decided to break the help website 1.2 into
multiple help pages in order of priority. First up is the sound-search
page describing how to work through data via google spreadsheets and the
listen page. All items in this commit are text and associated image and
audio media. Anthony, please add the widget text required to enable
playing of audio. Others in the lab are also very welcome to provide
input!!!!!!!