Adjustments to figure 2
I noticed this image is very blurry. This became particularly obvious
after laoding a new Fig 2 with arrow corrections and a higher res. I
have tried to replace Figure 1 in the same way but it continues to be
grainy when I committed to master (event the captions from ppt that
aren't in Fig 1). Rather than commit to the master again, I thought I
would create a branch so that @atruskie or @cofiem can have a look to
see if there is any way for me to improve the image quality to be
comparable to Fig 1.